requirements.txt
absl-py                       2.3.0
asgiref                       3.8.1
astunparse                    1.6.3
beautifulsoup4                4.13.4
blinker                       1.9.0
cachetools                    5.5.2
certifi                       2025.4.26
charset-normalizer            3.4.2
click                         8.2.1
colorama                      0.4.6
deepface                      0.0.93
Django                        5.2.1
djangorestframework           3.16.0
djangorestframework_simplejwt 5.5.0
drf-yasg                      1.21.10
filelock                      3.18.0
fire                          0.7.0
Flask                         3.1.1
flask-cors                    6.0.0
flatbuffers                   25.2.10
gast                          0.4.0
gdown                         5.2.0
google-auth                   2.40.2
google-auth-oauthlib          1.0.0
google-pasta                  0.2.0
grpcio                        1.71.0
gunicorn                      23.0.0
h5py                          3.13.0
idna                          3.10
inflection                    0.5.1
itsdangerous                  2.2.0
jax                           0.4.30
jaxlib                        0.4.30
Jinja2                        3.1.6
joblib                        1.5.1
keras                         2.12.0
libclang                      18.1.1
lz4                           4.4.4
Markdown                      3.8
MarkupSafe                    3.0.2
ml_dtypes                     0.5.1
mtcnn                         1.0.0
numpy                         1.23.5
oauthlib                      3.2.2
opencv-python                 4.11.0.86
opt_einsum                    3.4.0
packaging                     25.0
pandas                        2.2.3
pillow                        11.2.1
pip                           25.1.1
protobuf                      4.25.7
pyasn1                        0.6.1
pyasn1_modules                0.4.2
PyJWT                         2.9.0
PySocks                       1.7.1
python-dateutil               2.9.0.post0
pytz                          2025.2
PyYAML                        6.0.2
requests                      2.32.3
requests-oauthlib             2.0.0
retina-face                   0.0.17
rsa                           4.9.1
scipy                         1.15.3
setuptools                    58.1.0
six                           1.17.0
soupsieve                     2.7
sqlparse                      0.5.3
tensorboard                   2.12.3
tensorboard-data-server       0.7.2
tensorflow                    2.12.0
tensorflow-estimator          2.12.0
tensorflow-intel              2.12.0
tensorflow-io-gcs-filesystem  0.31.0
termcolor                     3.1.0
tqdm                          4.67.1
typing_extensions             4.13.2
tzdata                        2025.2
uritemplate                   4.1.1
urllib3                       2.4.0
Werkzeug                      3.1.3
wheel                         0.45.1
wrapt                         1.14.1